transfor the project to more modulize
